+++ lib/credo/check.ex
@@ -81,8 +81,11 @@ defmodule Credo.Check do
 
   @doc """
   Returns the base priority for the check.
+  
+  This can be one of `:higher`, `:high`, `:normal`, `:low` or `:ignore` 
+  (technically it can also be  or an integer, but these are internal representations although that is not recommended).
   """
-  @callback base_priority() :: integer
+  @callback base_priority() :: :higher | :high | :normal | :low | :ignore | integer
 
   @doc """
   Returns the category for the check.